`move compare_single_FBS_against_remote` to validation.py
